Move General class out of Program.cs
